+++ src/RecordTranslate/NullRecord.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Efabrica\Translatte\Record;
+
+class NullRecord implements RecordInterface
+{
+    public function save(string $message): void
+    {
+        // Nothing to do
+    }
+}