+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## [Unreleased]
 
-
+### Added
+- Record interface for translate record for statistic
 
 ## [1.5.0] - 2025-03-24
 ### Added